+++ exporters/otlp/otlp.go
@@ -21,7 +21,6 @@ import (
 	"context"
 	"errors"
 	"sync"
-	"time"
 
 	"google.golang.org/grpc"
 
@@ -30,14 +29,12 @@ import (
 	coltracepb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/collector/trace/v1"
 	"go.opentelemetry.io/otel/exporters/otlp/internal/transform"
 	"go.opentelemetry.io/otel/metric"
-	"go.opentelemetry.io/otel/propagation"
 	metricsdk "go.opentelemetry.io/otel/sdk/export/metric"
 	"go.opentelemetry.io/otel/sdk/export/metric/aggregation"
 	tracesdk "go.opentelemetry.io/otel/sdk/export/trace"
 	"go.opentelemetry.io/otel/sdk/metric/controller/push"
 	"go.opentelemetry.io/otel/sdk/metric/processor/basic"
 	"go.opentelemetry.io/otel/sdk/metric/selector/simple"
-	"go.opentelemetry.io/otel/sdk/resource"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 )
 
@@ -240,47 +237,43 @@ func (e *Exporter) uploadTraces(ctx context.Context, sdl []*tracesdk.SpanData) e
 	return err
 }
 
-func NewExportPipeline(ctx context.Context, exportOpts []ExporterOption,
-	resOpts []resource.Option) (*Exporter, *sdktrace.TracerProvider, *push.Controller, error) {
+// NewExportPipeline sets up a complete export pipeline
+// with the recommended setup for trace provider.
+func NewExportPipeline(ctx context.Context, exporterOpts ...ExporterOption) (*Exporter,
+	*sdktrace.TracerProvider, *push.Controller, error) {
 
-	exp, err := NewExporter(ctx, exportOpts...)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	res, err := resource.New(ctx, resOpts...)
+	exp, err := NewExporter(ctx, exporterOpts...)
 	if err != nil {
 		return nil, nil, nil, err
 	}
 
 	bsp := sdktrace.NewBatchSpanProcessor(exp)
 	tracerProvider := sdktrace.NewTracerProvider(
-		sdktrace.WithConfig(sdktrace.Config{DefaultSampler: sdktrace.AlwaysSample()}),
-		sdktrace.WithResource(res),
+		sdktrace.WithBatcher(exp),
 		sdktrace.WithSpanProcessor(bsp),
 	)
 
 	pusher := push.New(
 		basic.New(
-			simple.NewWithExactDistribution(),
+			simple.NewWithInexpensiveDistribution(),
 			exp,
 		),
 		exp,
-		push.WithPeriod(7*time.Second),
 	)
 
 	return exp, tracerProvider, pusher, nil
 }
 
-func InstallNewPipeline(ctx context.Context, exportOpts []ExporterOption,
-	resOpts []resource.Option) (*Exporter, *sdktrace.TracerProvider, *push.Controller, error) {
+// InstallNewPipeline instantiates a NewExportPipeline with the
+// recommended configuration and registers it globally.
+func InstallNewPipeline(ctx context.Context, exporterOpts ...ExporterOption) (*Exporter,
+	*sdktrace.TracerProvider, *push.Controller, error) {
 
-	exp, tp, pusher, err := NewExportPipeline(ctx, exportOpts, resOpts)
+	exp, tp, pusher, err := NewExportPipeline(ctx, exporterOpts...)
 	if err != nil {
 		return nil, nil, nil, err
 	}
 
-	otel.SetTextMapPropagator(propagation.TraceContext{})
 	otel.SetTracerProvider(tp)
 	otel.SetMeterProvider(pusher.MeterProvider())
 	pusher.Start()